pyproject.toml
[tool.poetry]
name = "flask-celery-project"
version = "0.1.0"
description = ""
authors = ["YuZhanHuang <bigjoe1992111@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "2.0.1"
celery = "5.1.2"
redis = "3.5.3"
flower = "1.0.0"
Flask-Migrate = "3.1.0"
Flask-SQLAlchemy = "2.5.1"
Flask-CeleryExt = "0.3.4"
psycopg2-binary = "2.9.1"
watchgod = "0.7"
Flask-WTF = "0.15.1"
requests = "2.26.0"
Flask-SocketIO = "5.0.1"
pytest = "6.2.4"
pytest-flask = "1.2.0"
factory-boy = "3.2.0"
pytest-factoryboy = "2.1.0"
pytest-cov = "2.12.1"
Pillow = "8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
